Generate Todolist project using create-react-app
This project should be used to serve as an intro to working with
React from the more standard way of working with React projects.
It may also serve as an introduction to state management tools
like Redux, but I am currently undecided on whether Redux should
be brought in at this point.